% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{build_matrix}
\alias{build_matrix}
\title{Create a matrix from columns of a tidy data frame}
\usage{
build_matrix(.data, rows, cols, elts, fill = NULL)
}
\arguments{
\item{.data}{A data frame with columns \code{rows}, \code{cols}, and \code{elts}.}

\item{rows}{Column that will become the rownames of the matrix.}

\item{cols}{Column that will become the colnames of the matrix.}

\item{elts}{Column that will become the matrix elements.}

\item{fill}{Value to use for missing elements.}
}
\description{
Note: Setting \code{fill} will replace NAs and NaNs, along with elements
corresponding to missing rows, with the value of \code{fill}.
}
