% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\alias{calibrate.matrix}
\alias{calibrate.otu_table}
\alias{calibrate.phyloseq}
\title{Calibrate a relative-abundance matrix by a bias vector}
\usage{
calibrate(observed, bias, ...)

\method{calibrate}{matrix}(observed, bias, margin, norm = "close")

\method{calibrate}{otu_table}(observed, bias, norm = "close")

\method{calibrate}{phyloseq}(observed, bias, norm = "close")
}
\arguments{
\item{observed}{An abundance matrix or phyloseq object containing one}

\item{bias}{A numeric vector of relative efficiencies}

\item{margin}{Matrix margin that corresponds to observations (samples);
\code{1} for rows, \code{2} for columns}

\item{norm}{String specifying how to normalize the calibrated observations;
see Details.}
}
\description{
Calibration via the simple deterministic procedure described in McLaren,
Willis, and Callahan (2019), simply involved dividing the observed vector of
relative abundances by the estimated bias vector and (optionally)
normalizing the result to sum to 1 or some other chosen value.
}
\details{
Normalization options specified by \code{norm}:
\itemize{
\item "close": Divide the calibrated abundance vector by its sum, so that it
sums to 1
\item "keep": Keep the same total abundance as the original observation
\item "none": Return the calibrated abundances without any normalization
}

If \code{bias} is named, then the names must agree with the feature (taxa) names
in \code{observed} and will be used to reorder \code{bias} to match the feature order
in \code{observed}.
}
\seealso{
\code{\link[=perturb]{perturb()}} \code{\link[=estimate_bias]{estimate_bias()}}
}
