% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{clr}
\alias{clr}
\title{Centered log-ratio transform of x.}
\usage{
clr(x, base = exp(1), na_rm = FALSE)
}
\arguments{
\item{x}{Vector of abundances.}

\item{base}{Base for logarithm}
}
\description{
Centered log-ratio transform of x.
}
