% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratios.R
\name{compute_ratios}
\alias{compute_ratios}
\title{Compute taxon ratios from a tidy data frame}
\usage{
compute_ratios(.data, group_vars = c("Sample"),
  comp_vars = setdiff(names(dplyr::select_if(.data, is.numeric)),
  group_vars), drop = TRUE)
}
\arguments{
\item{drop}{Whether to drop the individual taxon quantities}
}
\description{
Assumes the taxon variable name is \code{Taxon}
}
