% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias-helpers.R
\name{cooccurrence}
\alias{cooccurrence}
\alias{cooccurrence_matrix}
\alias{cooccurrence_network}
\alias{cooccurrence_components}
\title{Taxon co-occurrence network}
\usage{
cooccurrence_matrix(mat, all = TRUE)

cooccurrence_network(mat, all = TRUE)

cooccurrence_components(mat, all = TRUE, enframe = FALSE)
}
\arguments{
\item{mat}{A compositional error matrix}

\item{all}{Whether to include taxa that aren't in any multi-taxon samples}

\item{enframe}{Whether to "enframe" the returned vector}
}
\description{
Functions for probing the taxon co-occurrence network, used for checking if
the bias estimated by the \code{center()} function is fully-determined.
}
\details{
The edges are weighted by the number of samples the pair of taxa co-occur in
and so provide some information about the precision of the pairwise bias
estimate.
}
\section{Functions}{
\itemize{
\item \code{cooccurrence_matrix}: Adjacency matrix of the co-occurrence network

\item \code{cooccurrence_network}: Co-occurrence network as an \code{igraph} object

\item \code{cooccurrence_components}: Connected components of the co-occurrence network
}}

