% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_bias}
\alias{estimate_bias}
\alias{estimate_bias.matrix}
\alias{estimate_bias.otu_table}
\alias{estimate_bias.phyloseq}
\title{Estimate bias from control measurements}
\usage{
estimate_bias(observed, actual, ...)

\method{estimate_bias}{matrix}(observed, actual, margin, ...)

\method{estimate_bias}{otu_table}(observed, actual, ...)

\method{estimate_bias}{phyloseq}(observed, actual, ...)
}
\arguments{
\item{observed}{Abundance matrix of observed compositions}

\item{actual}{Abundance matrix of actual or reference compositions for the
same samples and taxa in \code{observed}}

\item{...}{Arguments passed to \code{\link[=center]{center()}}}

\item{margin}{Matrix margin that corresponds to observations (samples);
\code{1} for rows, \code{2} for columns}
}
\description{
Estimate bias using the compositional least-squares approach described in
McLaren, Willis, and Callahan (2019).
}
\details{
Name requirements for \code{observed} and \code{actual}: The row and column names (for
matrices) or taxa and sample names (for phyloseq objects) must match, but
can be in different orders.
}
\seealso{
\code{\link[=center]{center()}} \code{\link[=calibrate]{calibrate()}}
}
