% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositional-mean.R
\name{log_center_rss}
\alias{log_center_rss}
\title{Compute the log center by numerical optimization}
\usage{
log_center_rss(mat, weights = rep(1, nrow(mat)), bound = 10)
}
\arguments{
\item{bound}{Single number giving the lower and upper bound on the alr
efficiencies ("rss" only).}
}
\description{
Compute the log center by numerical optimization
}
