% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacal-package.R
\docType{package}
\name{metacal-package}
\alias{metacal}
\alias{metacal-package}
\title{metacal: Bias estimation and calibration for metagenomics experiments}
\description{
The metacal package provides tools for bias estimation and calibration in marker-gene and metagenomics sequencing experiments, using the methods developed in https://elifesciences.org/articles/46923.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mikemc.github.io/metacal}
  \item \url{https://github.com/mikemc/metacal}
  \item Report bugs at \url{https://github.com/mikemc/metacal/issues}
}

}
\author{
\strong{Maintainer}: Michael McLaren \email{m.mclaren42@gmail.com} (\href{https://orcid.org/0000-0003-1575-473X}{ORCID})

}
\keyword{internal}
