% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratios.R
\name{pairwise_ratios}
\alias{pairwise_ratios}
\alias{pairwise_ratios.default}
\alias{pairwise_ratios.matrix}
\alias{pairwise_ratios.otu_table}
\alias{pairwise_ratios.phyloseq}
\title{Pairwise ratios of vector elements and matrix rows or columns}
\usage{
pairwise_ratios(x, ..., f = `/`, filter = TRUE, set_names = TRUE, sep = ":")

pairwise_ratios.default(x, f = `/`, filter = TRUE, set_names = TRUE, sep = ":")

pairwise_ratios.matrix(
  x,
  margin,
  f = `/`,
  filter = TRUE,
  set_names = TRUE,
  sep = ":"
)

pairwise_ratios.otu_table(
  x,
  margin = "taxa",
  f = `/`,
  filter = TRUE,
  sep = ":"
)

pairwise_ratios.phyloseq(x, margin = "taxa", f = `/`, filter = TRUE, sep = ":")
}
\arguments{
\item{x}{A vector or matrix}

\item{f}{Vectorized binary function applied to numerator and denominator}

\item{filter}{Whether to filter out redundant pairs}

\item{set_names}{Whether to set names for the pairs}

\item{sep}{Seperator for pair names}

\item{margin}{Margin (1 = rows, 2 = cols) of matrix x to compute ratios of}
}
\description{
For a matrix, computes the ratios of the supplied margin (1 = rows, 2 =
cols). Otherwise, computes the ratios of the elements.
}
\examples{
  mat <- seq(10) |>
    matrix(nrow = 2)
  rownames(mat) <- paste0('r', seq(2))
  colnames(mat) <- paste0('c', seq(5))
  
  pairwise_ratios(mat, 2)
  pairwise_ratios(mat[1,])
}
