% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{perturb}
\alias{perturb}
\alias{perturb.matrix}
\alias{perturb.otu_table}
\alias{perturb.phyloseq}
\title{Compositionally perturb a relative-abundance matrix}
\usage{
perturb(x, y, ...)

\method{perturb}{matrix}(x, y, margin, norm = "close")

\method{perturb}{otu_table}(x, y, norm = "close")

\method{perturb}{phyloseq}(x, y, norm = "close")
}
\arguments{
\item{x}{An abundance matrix or phyloseq object containing one}

\item{y}{A numeric vector with which to perturb the observations in x}

\item{margin}{Matrix margin that corresponds to observations (samples);
\code{1} for rows, \code{2} for columns}

\item{norm}{String specifying how to normalize the perturbed observations;
see Details.}
}
\description{
The perturbation of two compositional vectors \code{v} and \code{y} is given (up to
compositional equivalence) by the elementwise product \code{v * y}.
}
\details{
Let \code{w = v * y}, where \code{v} is a row or column of \code{x}. The normalization
options specified by the \code{norm} argument are
\itemize{
\item "close" Return the compositional closure of w into the simplex,
\code{w / sum(w)}
\item "keep" Keep the same total abundance as the original vector by returning
\code{w * sum(v) / sum(w)}
\item "none" Return \code{w} without any normalization
}
}
