% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositional-mean.R
\name{proj_mat}
\alias{proj_mat}
\title{Projection matrix P_M}
\usage{
proj_mat(K, M = c())
}
\arguments{
\item{K}{number of elements (taxa)}

\item{M}{set of missing elements (default is none missing)}
}
\description{
Used internally for the "proj" method of computing the compositional mean.
See vandenBoogaart2006 and Bren2008.
}
