% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{xydist}
\alias{xydist}
\title{Distance or dissimilarity between relative abundance vectors x and y}
\usage{
xydist(x, y, method = "aitchison", trim = FALSE)
}
\arguments{
\item{method}{Distance/dissimilarity measure.}

\item{trim}{Should x and y be reduced to their common positive elements
before computing the Aitchison distance (otherwise, the distance will be
Inf).

method == "aitchison" -> Aitchison distance
method == "bray" -> Bray-Curtis dissimilarity between x and y. Note,
converts x and y to proportions before computing.}
}
\description{
Distance or dissimilarity between relative abundance vectors x and y
}
