% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositional-mean.R
\name{bootrep_center}
\alias{bootrep_center}
\title{Generate bootstrap replicates of the sample center}
\usage{
bootrep_center(.data, R = 4000, N = nrow(.data), method = "proj",
  dist = "dirichlet", in_scale = "linear", out_scale = "linear",
  denom = NULL)
}
\arguments{
\item{.data}{\itemize{
\item all-numeric data frame or matrix with taxa as columns
}}

\item{R}{Number of bootstrap replicates}

\item{N}{Number of trials for multinomial resampling}

\item{method}{\itemize{
\item "proj", "gm", or "rss"
}}

\item{dist}{"dirichlet" or "multinomial" resampling}

\item{in_scale}{\itemize{
\item "linear" (default) or "log"
}}

\item{out_scale}{\itemize{
\item "linear" (default) or "log"
}}

\item{denom}{\itemize{
\item Taxa to use in the denominator; if NULL, use all taxa.
}}
}
\description{
Generate bootstrap replicates of the sample center
}
