% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositional-mean.R
\name{bootrep_center}
\alias{bootrep_center}
\title{Generate bootstrap replicates of the sample center}
\usage{
bootrep_center(.data, R = 4000, N = nrow(.data), method = "proj",
  dist = "dirichlet", in_scale = "linear", out_scale = "linear",
  denom = NULL)
}
\description{
Generate bootstrap replicates of the sample center
}
