% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{clr}
\alias{clr}
\title{Compute the centered log-ratio transform of x}
\usage{
clr(x, base = exp(1), na.rm = FALSE)
}
\arguments{
\item{x}{Vector of abundances.}

\item{base}{Base for logarithm}
}
\description{
Compute the centered log-ratio transform of x
}
