% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_bias}
\alias{estimate_bias}
\alias{estimate_bias.matrix}
\alias{estimate_bias.otu_table}
\alias{estimate_bias.phyloseq}
\title{Estimate bias from control measurements}
\usage{
estimate_bias(observed, actual, ...)

\method{estimate_bias}{matrix}(observed, actual, margin, ...)

\method{estimate_bias}{otu_table}(observed, actual, ...)

\method{estimate_bias}{phyloseq}(observed, actual, ...)
}
\arguments{
\item{observed}{Abundance matrix of observed compositions}

\item{actual}{Abundance matrix of actual or reference compositions for the
same samples and taxa in \code{observed}}

\item{...}{Arguments passed to methods.}

\item{margin}{Matrix margin that corresponds to observations (samples);
\code{1} for rows, \code{2} for columns}
}
\value{
A \code{mc_bias_fit} object with \code{\link[=coef]{coef()}}, \code{\link[=fitted]{fitted()}}, and \code{\link[=residuals]{residuals()}}
methods.
}
\description{
Estimate bias using the compositional least-squares approach described in
McLaren, Willis, and Callahan (2019).
}
\details{
Bias is estimated by applying \code{\link[=center]{center()}} to the compositional error matrix
defined by \code{observed/actual}, which requires that \code{observed} and \code{actual}
are non-zero for the same sample-feature pairs. For convenience, this
function will automatically set values in \code{observed} to 0 whose
corresponding entries are 0 in \code{actual}, but it is up to you to replace 0
values in \code{observed} with a non-zero value (such as a pseudocount).

Name requirements for \code{observed} and \code{actual}: The row and column names (for
matrices) or taxa and sample names (for phyloseq objects) must match, but
can be in different orders.
}
\seealso{
\code{\link[=center]{center()}} \code{\link[=calibrate]{calibrate()}}
}
