% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_bias}
\alias{estimate_bias}
\alias{estimate_bias.matrix}
\alias{estimate_bias.otu_table}
\alias{estimate_bias.phyloseq}
\title{Estimate bias from control measurements}
\usage{
estimate_bias(observed, actual, ...)

\method{estimate_bias}{matrix}(observed, actual, margin, boot = FALSE, times = 1000)

\method{estimate_bias}{otu_table}(observed, actual, ...)

\method{estimate_bias}{phyloseq}(observed, actual, ...)
}
\arguments{
\item{observed}{Abundance matrix of observed compositions.}

\item{actual}{Abundance matrix of actual or reference compositions for the
same samples and taxa in \code{observed}.}

\item{...}{Arguments passed to the matrix method.}

\item{margin}{Matrix margin that corresponds to observations (samples);
\code{1} for rows, \code{2} for columns.}

\item{boot}{Whether to perform bootstrapping.}

\item{times}{Number of bootstrap replicates.}
}
\value{
A \code{mc_bias_fit} object with \code{\link[=coef]{coef()}}, \code{\link[=fitted]{fitted()}}, \code{\link[=residuals]{residuals()}}, and
\code{\link[=summary]{summary()}} methods.
}
\description{
Estimate bias using the compositional least-squares approach described in
McLaren, Willis, and Callahan (2019).
}
\details{
Bias is estimated by applying \code{\link[=center]{center()}} to the compositional error matrix
defined by \code{observed/actual}, which requires that \code{observed} and \code{actual}
are non-zero for the same sample-taxa pairs. For convenience, this
function will automatically set values in \code{observed} to 0 whose
corresponding entries are 0 in \code{actual}, but it is up to you to replace 0
values in \code{observed} with a non-zero value (such as a pseudocount).

Name requirements for \code{observed} and \code{actual}: The row and column names (for
matrices) or taxa and sample names (for phyloseq objects) must match, but
can be in different orders.
}
\examples{
# Load data from the cellular mock communities of Brooks et al 2015
dr <- system.file("extdata", package = "metacal")
list.files(dr)
actual <- file.path(dr, "brooks2015-actual.csv") |>
  read.csv(row.names = "Sample") |>
  as("matrix")
observed <- file.path(dr, "brooks2015-observed.csv") |>
  read.csv(row.names = "Sample") |>
  subset(select = - Other) |>
  as("matrix")
sam <- file.path(dr, "brooks2015-sample-data.csv") |> read.csv()

# Estimate bias with bootstrapping for error estimation
mc_fit <- estimate_bias(observed, actual, margin = 1, boot = TRUE)
summary(mc_fit)
}
\seealso{
\code{\link[=center]{center()}} \code{\link[=calibrate]{calibrate()}}
}
