% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_efficiency.R
\name{mean_efficiency}
\alias{mean_efficiency}
\alias{mean_efficiency.mc_bias_fit}
\alias{mean_efficiency.numeric}
\title{Compute sample mean efficiencies}
\usage{
mean_efficiency(bias, actual, observed, margin)

\method{mean_efficiency}{mc_bias_fit}(bias, actual = NULL, observed = NULL, margin = NULL)

\method{mean_efficiency}{numeric}(bias, actual = NULL, observed = NULL, margin)
}
\arguments{
\item{bias}{A numeric vector of relative efficiencies or an object of class
'mc_bias_fit' (from which efficiencies will be extracted)}

\item{actual}{A matrix or NULL}

\item{observed}{A matrix or NULL}

\item{margin}{The margin containing the samples/observations}
}
\description{
Given an \code{mc_bias_fit} object or a vector of taxa efficiencies, compute the
mean efficiency in the control samples in the \code{mc_bias_fit} object or a set
of user-supplied samples. This function currently only works on matrices,
not phyloseq objects.
}
\details{
If \code{bias} is an \code{mc_bias_fit} object and \code{actual} and \code{observed} are both
missing, then the estimated mean efficiency for the control samples used in
the \code{mc_bias_fit} object will be computed. New samples should be given in
either \code{actual} or \code{observed}, depending on whether they are nominal or
calibrated abundances or observed (i.e. uncalibrated) data; observed data
will be calibrated first.

The mean efficiencies in a sample with is equal to \verb{weighted.mean(bias, x))}
where \code{bias} is the vector of taxa efficiencies and \code{x} is the vector of
taxa proportions.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
\code{\link[=estimate_bias]{estimate_bias()}}
}
