% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{pw_ratios_matrix}
\alias{pw_ratios_matrix}
\title{Pairwise ratios of taxa from matrix of multiple samples}
\usage{
pw_ratios_matrix(x, na_rm = FALSE)
}
\description{
Pairwise ratios of taxa from matrix of multiple samples
}
